+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Add import_records arg to `bp.load_project()` [x]
 * Improve reserved aminoacid characters [x]
 * Add `SeqFile.max_seq` and `.min_seq` properties [x]
+* Patch `Project` deserializer to improve BioProvDocument creation [x]
 
 ### v0.1.22
 * Simplify `bp.load_project()` function [x]