+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Add `SeqFile.max_seq` and `.min_seq` properties [x]
 * Patch `Project` deserializer to improve BioProvDocument creation [x]
 * Make shorter Environment hashes [x]
+* Improve Project `__repr__` [x]
 
 ### v0.1.22
 * Simplify `bp.load_project()` function [x]